refactor(cli): [TASK-122] fix #122 - deduplicate detectInstalledPackageManager
- Deduplicate detectInstalledPackageManager() in CLI
  Consolidated two separate implementations into a single robust function
  in packageManagerDetect.js that checks npm_config_user_agent env var
  first, then falls back to testing actual PATH availability via
  execFileSync. Exported getPreferredPackageManager and
  setPreferredPackageManager from the shared module. Updated
  updateHandler.js to import from the shared source.
