+++ packages/cli/src/utils/packageManagerDetect.js
@@ -1,60 +1,104 @@
-// Detect package manager from environment and process arguments
-// Used internally for auto-install, not for user prompting
-function detectFromEnv() {
-  const userAgent = process.env.npm_config_user_agent;
-  if (!userAgent) return null;
+import { execFileSync } from "child_process";
+import fs from "fs";
+import os from "os";
+import path from "path";
 
-  if (userAgent.includes('bun')) return 'bun';
-  if (userAgent.includes('yarn')) return 'yarn';
-  if (userAgent.includes('pnpm')) return 'pnpm';
-  if (userAgent.includes('npm')) return 'npm';
-  
-  return null;
+const _DOCUBOOK_DIR = path.join(os.homedir(), ".docubook");
+const _PREFERENCES_FILE = path.join(_DOCUBOOK_DIR, "cli-config.json");
+
+function _readPreferences() {
+  try {
+    const raw = fs.readFileSync(_PREFERENCES_FILE, "utf8");
+    return JSON.parse(raw || "{}");
+  } catch {
+    return {};
+  }
 }
 
-// Detect from process argv (how the CLI was invoked)
-function detectFromArgv() {
-  const argv1 = process.argv[1] || '';
-  
-  if (argv1.includes('bunx') || argv1.includes('bun')) return 'bun';
-  if (argv1.includes('pnpm')) return 'pnpm';
-  if (argv1.includes('yarn')) return 'yarn';
-  if (argv1.includes('npx') || argv1.includes('npm')) return 'npm';
-  
-  return null;
+function _writePreferences(obj) {
+  try {
+    if (!fs.existsSync(_DOCUBOOK_DIR)) {
+      fs.mkdirSync(_DOCUBOOK_DIR, { recursive: true, mode: 0o700 });
+    }
+    fs.writeFileSync(_PREFERENCES_FILE, JSON.stringify(obj, null, 2), { mode: 0o600 });
+  } catch {
+    // non-fatal
+  }
 }
 
 /**
- * Detect which package manager invoked the CLI
- * Used internally for auto-install dependency resolution
- * @returns {string} Package manager name (default: 'npm')
+ * Detect which package manager is available.
+ * Checks npm_config_user_agent env var first, then falls back to testing
+ * actual availability in PATH via execFileSync.
+ * @returns {string} Package manager name ('npm' | 'pnpm' | 'yarn' | 'bun')
  */
 export function detectInstalledPackageManager() {
-  const detected = detectFromEnv() || detectFromArgv();
-  return detected || 'npm';
+  try {
+    const userAgent = process.env.npm_config_user_agent || "";
+    if (userAgent.includes("npm")) return "npm";
+    if (userAgent.includes("pnpm")) return "pnpm";
+    if (userAgent.includes("yarn")) return "yarn";
+    if (userAgent.includes("bun")) return "bun";
+
+    // Fallback: check what's available in PATH
+    const candidates = ["npm", "pnpm", "yarn", "bun"];
+    for (const pm of candidates) {
+      try {
+        execFileSync(pm, ["--version"], { stdio: "ignore" });
+        return pm;
+      } catch {
+        // try next
+      }
+    }
+
+    return "npm";
+  } catch {
+    return "npm";
+  }
+}
+
+/**
+ * Get preferred package manager from persisted config, falling back to detection.
+ * @returns {string} Package manager name
+ */
+export function getPreferredPackageManager() {
+  const prefs = _readPreferences();
+  if (prefs.packageManager) return prefs.packageManager;
+  return detectInstalledPackageManager();
+}
+
+/**
+ * Persist the preferred package manager to ~/.docubook/cli-config.json
+ * @param {string} pm - Package manager name
+ */
+export function setPreferredPackageManager(pm) {
+  if (!pm) return;
+  const prefs = _readPreferences();
+  prefs.packageManager = pm;
+  _writePreferences(prefs);
 }
 
 // Get package manager info (static, cached)
 const PACKAGE_MANAGER_INFO = {
   npm: {
-    name: 'npm',
-    installCmd: 'npm install',
-    devCmd: 'npm run dev',
+    name: "npm",
+    installCmd: "npm install",
+    devCmd: "npm run dev",
   },
   yarn: {
-    name: 'yarn',
-    installCmd: 'yarn install',
-    devCmd: 'yarn dev',
+    name: "yarn",
+    installCmd: "yarn install",
+    devCmd: "yarn dev",
   },
   pnpm: {
-    name: 'pnpm',
-    installCmd: 'pnpm install',
-    devCmd: 'pnpm dev',
+    name: "pnpm",
+    installCmd: "pnpm install",
+    devCmd: "pnpm dev",
   },
   bun: {
-    name: 'bun',
-    installCmd: 'bun install',
-    devCmd: 'bun run dev',
+    name: "bun",
+    installCmd: "bun install",
+    devCmd: "bun run dev",
   },
 };
